- Added KrnStatMemory() function - Changed KrnAllocPages() specification, added support for allocating regions with given addresses
git-svn-id: https://svn.aros.org/svn/aros/trunk@35843 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/kernel/freepages.c
@@ -17,7 +17,7 @@ AROS_LH2(void, KrnFreePages,
 
 /*  SYNOPSIS */
 	AROS_LHA(void *, addr, A0),
-	AROS_LHA(uint32_t, length, D0),
+	AROS_LHA(uintptr_t, length, D0),
 
 /*  LOCATION */
 	struct KernelBase *, KernelBase, 28, Kernel)
